The data file now has mutual exclusion through a lock file thanks to locklib FriendlyConfig can now be set to not write the file if not found cleardata will now clear the configured datafile, and can be passed an argument to an alternate file

+++ FriendlyConfig.py
@@ -4,7 +4,7 @@ from ConfigParser import RawConfigParser
 #and will create the config file if it doesn't exist.
 #getmain* methods are like RawConfigParser.get* methods.  Arguments are option[, section] where section defaults to mainSection.
 class FriendlyConfig(RawConfigParser):
-	def __init__(self, filename, defaults={}, mainSection="main"):
+	def __init__(self, filename, defaults={}, mainSection="main", write=True):
 		#RawConfigParser is an old-style class
 		RawConfigParser.__init__(self, defaults)
 		self.filename = filename
@@ -22,7 +22,7 @@ class FriendlyConfig(RawConfigParser):
 		if not self.has_section(self.mainSection):
 			self.add_section(self.mainSection)
 		
-		if not configExists:
+		if not configExists and write:
 			self.write(open(self.filename, "w"))
 
 	def getmain(self, option, section=None, type=""):
